Fix failing unit tests (#77)
We merged in failing unit tests to avoid the initial "perfect is the
enemy of the good" situation. This patch fixes all the failing unit
tests so that we can rely on the continuous integration again.

(Note that the CI works only locally, with the most recent 
versions of `icontract` and `icontract-hypothesis`. Once they are
published, the "remote" CI will work as well.)