Fix ETHZ Eprog 2021 Ex. 8 Prob. 3 and 5 (#106)
This patch fixes a bug in the solution of ETHZ Eprog 2019 Ex. 8, Prob. 3
where statistics could not be computed on empty samples (when
the number of steps in the simulation was set to 0). We also propagate
the change down to the curated incorrect solutions.

Additionally, we fix the test for Prob. 5 of the same exercise to avoid
zero number of trials, for the same reason.